requirements.txt
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
pycodestyle==2.9.1
python-dotenv==0.20.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.11
